Implemented WordWrap with `set_wrap_mode` and Gtk.WrapMode.WORD to wrap text in chatbox applied to TextView().
Added Max Message Size /w Constant (size 200 subject to change) using the `set_max_length` method applied to Gtk.Entry().

Fixed Sizing of Chat Expander with method `set_min_content_height` applied to to ScrolledWindow() using constant (also subject to change).

Fixed Chat Entry and Button size by setting set_homogeneous False on the Gtk.Box (chat_entry_box) container.  Then when adding the components set Expand and Fill to True for the Entry() and False to the Gtk.Button().

Confirmed that the error regarding Gtk.AccelGroup is a sugar warning message and is not something we are responsible for fixing (probably an artifact of the Gtk3 conversion).